requirements.txt
Django==4.2
django-crispy-forms==2.0
crispy-bootstrap5==2023.10
python-markdown==3.4.4
django-filter==23.3